API/requirements.txt
fastapi 
uvicorn
Pillow 
opencv-python-headless 
numpy
inaSpeechSegmenter
moviepy
librosa
pandas
mediapipe